Modularize bincomp executable
